requirements.txt
openai>=1.40.0
Flask>=2.3
gunicorn>=21.2
python-dotenv>=1.0
